support one app expose multiple sockets
In some cases, on app has multiple sockets. Such as, a VDI system
has multiple VMs, each vm has a RDP service or SPICE service, each
service listen on a socket. So, a VDI app has to expose multiple
socket.

For `localExposedApps`, add `portRange` and `addressRange` to limit
the exposed sockets' range

For `remoteApps`，add `appSocket` to specify the `localSocket` access
which socket of the app.